# 03-data-persist-mongodb/docs/swagger.yaml
basePath: /
consumes:
- application/json
definitions:
  models.Recipe:
    properties:
      id:
        type: string
      ingredients:
        items:
          type: string
        type: array
      instructions:
        items:
          type: string
        type: array
      name:
        type: string
      published_at:
        type: string
      tags:
        items:
          type: string
        type: array
    type: object
host: localhost:9000
info:
  contact:
    name: Roshan Lamichhane
  description: This is a basic recipes API in go (Gin). A CRUD demonstration.
  title: Recipes API
  version: 1.0.0
paths:
  /recipes:
    get:
      consumes:
      - application/json
      description: List all the recipes
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.Recipe'
            type: array
        "500":
          description: Internal Server Error
      summary: List all the recipes
    post:
      consumes:
      - application/json
      description: Create a new recipe
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.Recipe'
        "400":
          description: Bad Request
      summary: Create a new recipe
  /recipes/id:
    delete:
      consumes:
      - application/json
      description: Delete a recipe
      produces:
      - application/json
      responses:
        "200":
          description: OK
        "404":
          description: Not Found
      summary: Delete a recipe
    put:
      consumes:
      - application/json
      description: Update a recipe
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Recipe'
        "404":
          description: Not Found
      summary: Update a recipe
produces:
- pplication/json
swagger: "2.0"